linux/kernel/common-config: support DFS radiations for ath(9|10)k drivers
This enables shady options because upstream and Atheros probably didn't fix the issue
and their idea of "certification" is quite extreme.

We follow OpenWRT here and apply those changes to reduce "WTFs/min" for users
wondering what is the difference between OpenWRT and NixOS for their APs.

This was tested on Qualcomm Atheros QCA986x/988x 802.11ac Wireless Network Adapter on a x86 machine
in the French regulatory domain on frequency 5600MHz, DFS was used successfully by hostapd
instead of a mind-boggling "device or resource busy" error which is what happens without those options.